+++ alsek/storage/backends/postgres/standard/backend.py
@@ -20,7 +20,7 @@ from alsek.storage.backends import Backend, LazyClient
 from alsek.storage.backends.postgres._utils import (
     validate_value_within_postgres_notification_size_limit,
 )
-from alsek.storage.backends.postgres.pubsub.standard import PostgresPubSubListener
+from alsek.storage.backends.postgres.standard._pubsub_listener import PostgresPubSubListener
 from alsek.storage.backends.postgres.tables import SCHEMA_NAME, Base
 from alsek.storage.backends.postgres.tables import KeyValue as KeyValueRecord
 from alsek.storage.backends.postgres.tables import KeyValueType
